backend/requirements.txt
Flask==3.1.0
Flask-CORS==6.0.0
torch==2.0.1
torchvision==0.15.2
Pillow==10.0.0
gunicorn==20.1.0
